+++ extension/CHANGELOG.md
@@ -16,7 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
 
+- Update leparisien integration to support other URL formats
 
 ## [0.7.0] - 2021-03-05